refactor: organize scenes
