+++ scenes/DeadScene.js
