+++ scenes/EncounterScene.js
